requirements.txt

asyncio==3.4.3
giphy-client==1.0.0
mysql-connector-python==8.0.32
Pillow==9.4.0
py-cord==2.3.2
python-dotenv==0.21.1
requests==2.28.2
ui==0.1.4
aiofiles==22.1.0